@@ -94,9 +102,13 @@ allOpen {
     annotation("no.nav.arbeidsplassen.importapi.Open")
 }
 
+kapt {
+    correctErrorTypes true
+}
+
 compileKotlin {
     kotlinOptions {
-        jvmTarget = "17"
+        jvmTarget = "21"
         //Will retain parameter names for Java reflection
         javaParameters = true
     }
@@ -104,7 +116,7 @@ compileKotlin {
 
 compileTestKotlin {
     kotlinOptions {
-        jvmTarget = "17"
+        jvmTarget = "21"
         javaParameters = true
     }
 }
