@@ -8,7 +8,7 @@ import io.micronaut.http.annotation.QueryValue
 import io.swagger.v3.oas.annotations.Hidden
 import no.nav.arbeidsplassen.importapi.dto.AdStateDTO
 import java.time.LocalDateTime
-import javax.annotation.security.PermitAll
+import jakarta.annotation.security.PermitAll
 
 @PermitAll
 @Controller("/internal/adstates")
